chore: support batch4 chains with deployments refactor
