+++ apps/lite/package.json
@@ -38,8 +38,8 @@
     "tailwind-merge": "^3.0.1",
     "tailwindcss-animate": "^1.0.7",
     "urql": "^4.2.2",
-    "viem": "~2.23.5",
-    "wagmi": "2.14.15"
+    "viem": "~2.29.1",
+    "wagmi": "2.15.2"
   },
   "devDependencies": {
     "@fleek-platform/cli": "^3.8.2",
