+++ apps/lite/src/app/dashboard/borrow-subpage.tsx
@@ -3,6 +3,7 @@ import { metaMorphoFactoryAbi } from "@morpho-org/uikit/assets/abis/meta-morpho-
 import { morphoAbi } from "@morpho-org/uikit/assets/abis/morpho";
 import useContractEvents from "@morpho-org/uikit/hooks/use-contract-events/use-contract-events";
 import { readAccrualVaults, readAccrualVaultsStateOverride } from "@morpho-org/uikit/lens/read-vaults";
+import { CORE_DEPLOYMENTS, getContractDeploymentInfo } from "@morpho-org/uikit/lib/deployments";
 import { restructure } from "@morpho-org/uikit/lib/restructure";
 import { getTokenSymbolURI, Token } from "@morpho-org/uikit/lib/utils";
 import { useMemo } from "react";
@@ -16,7 +17,6 @@ import { useMarkets } from "@/hooks/use-markets";
 import * as Merkl from "@/hooks/use-merkl-campaigns";
 import { useMerklOpportunities } from "@/hooks/use-merkl-opportunities";
 import { useTopNCurators } from "@/hooks/use-top-n-curators";
-import { CORE_DEPLOYMENTS, getContractDeploymentInfo } from "@/lib/constants";
 import { type DisplayableCurators, getDisplayableCurators } from "@/lib/curators";
 
 const STALE_TIME = 5 * 60 * 1000;
