+++ apps/lite/src/hooks/use-markets.ts
@@ -2,12 +2,11 @@ import { Market, MarketParams } from "@morpho-org/blue-sdk";
 import { adaptiveCurveIrmAbi } from "@morpho-org/uikit/assets/abis/adaptive-curve-irm";
 import { morphoAbi } from "@morpho-org/uikit/assets/abis/morpho";
 import { oracleAbi } from "@morpho-org/uikit/assets/abis/oracle";
+import { getContractDeploymentInfo } from "@morpho-org/uikit/lib/deployments";
 import { restructure } from "@morpho-org/uikit/lib/restructure";
 import { useMemo } from "react";
 import { Hex } from "viem";
 import { useReadContracts } from "wagmi";
-
-import { getContractDeploymentInfo } from "@/lib/constants";
 
 export function useMarkets({
   chainId,
