+++ packages/uikit/src/lib/chains/index.ts
@@ -1,0 +1,2 @@
+export { basecamp } from "./camp";
+export { hyperevm } from "./hyperevm";
